Rename kebab-case folder to camelCase
