+++ nodes/Anclap/actions/info/index.ts
@@ -2,11 +2,11 @@ import description from './description';
 import { getTransactions as transactions } from './transactions/execute';
 import { getTransaction as transaction } from './transaction/execute';
 import { getFeeAmount as fee } from './fee/execute';
-import { getTransferServer as transferServer } from './transfer-server/execute';
-import { getQuoteServer as quoteServer } from './quote-server/execute';
+import { getTransferServer as transferServer } from './transferServer/execute';
+import { getQuoteServer as quoteServer } from './quoteServer/execute';
 import { getPrices as prices } from './prices/execute';
 import { getPrice as price } from './price/execute';
-import { getFirmQuote as firmQuote } from './firm-quote/execute';
+import { getFirmQuote as firmQuote } from './firmQuote/execute';
 import { getQuote as quote } from './quote/execute';
 export {
 	transactions,
