+++ nodes/Anclap/actions/transactions/index.ts
@@ -1,10 +1,10 @@
 import description from './description';
 import { execute as deposit } from './deposit';
 import { execute as withdraw } from './withdraw';
-import { execute as depositInteractive } from './deposit-interactive';
-import { execute as withdrawInteractive } from './withdraw-interactive';
-import { execute as depositExchange } from './deposit-exchange';
-import { execute as withdrawExchange } from './withdraw-exchange';
+import { execute as depositInteractive } from './depositInteractive';
+import { execute as withdrawInteractive } from './withdrawInteractive';
+import { execute as depositExchange } from './depositExchange';
+import { execute as withdrawExchange } from './withdrawExchange';
 export {
 	deposit,
 	withdraw,
